fix: ファイル名マスク処理を共通部分に移動
