+++ hooq-macros/Cargo.toml
@@ -2,7 +2,7 @@
 name = "hooq-macros"
 version = "0.1.0"
 edition = "2024"
-rust-version = "1.88.0"
+rust-version = { workspace = true }
 
 [lib]
 proc-macro = true
@@ -16,9 +16,9 @@ toml = "0.9.5"
 
 [dev-dependencies]
 futures = "0.3.31"
-glob = "0.3.3"
 hooq = { path = "../hooq" }
 macrotest = "1.2.0"
 serde_json = "1.0.143"
 tokio = { version = "1.47.1", features = ["full"] }
 util_macros = { path = "tests/util_macros" }
+test-helpers = { path = "../test-helpers" }
