+++ hooq-macros/tests/special/meta_var/mod.rs
@@ -2,13 +2,16 @@ mod abs_path;
 mod bindings;
 mod hooq_meta;
 
+use test_helpers::MaskMode::*;
+use test_helpers::mask_project_root;
+
 #[test]
 fn test_special_meta_var() {
-    crate::mask_project_root("tests/special/meta_var", crate::UnMask);
+    mask_project_root("tests/special/meta_var", UnMask);
 
     macrotest::expand_args("tests/special/meta_var/abs_path.rs", &["--ugly"]);
     macrotest::expand_args("tests/special/meta_var/bindings.rs", &["--ugly"]);
     macrotest::expand_args("tests/special/meta_var/hooq_meta.rs", &["--ugly"]);
 
-    crate::mask_project_root("tests/special/meta_var", crate::Mask);
+    mask_project_root("tests/special/meta_var", Mask);
 }
