+++ test-helpers/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "test-helpers"
+version = "0.1.0"
+edition = "2024"
+rust-version.workspace = true
+
+[dependencies]
+glob = "0.3.3"
